+++ yorm/base/mappable.py
@@ -16,49 +16,38 @@ class Mappable(metaclass=abc.ABCMeta):
 
     def __getattribute__(self, name):
         if name == MAPPER:
+            # avoid infinite recursion (attribute requested in this function)
             return object.__getattribute__(self, name)
-
         mapper = getattr(self, MAPPER, None)
 
+        # Get the attribute's current value
         try:
             value = object.__getattribute__(self, name)
         except AttributeError as exc:
-            if mapper:
-                mapper.fetch()
-                value = object.__getattribute__(self, name)
-            else:
+            missing = True
+            if not mapper:
                 raise exc from None
         else:
-            if mapper and name in mapper.attrs:
-                mapper.fetch()
-                value = object.__getattribute__(self, name)
+            missing = False
+
+        # Fetch a new value from disk if the attribute is mapped or missing
+        if mapper and (missing or name in mapper.attrs):
+            mapper.fetch()
+            value = object.__getattribute__(self, name)
 
         return value
 
     def __setattr__(self, name, value):
         mapper = getattr(self, MAPPER, None)
 
+        # Convert the value to the mapped type
         if mapper and name in mapper.attrs:
             converter = mapper.attrs[name]
             value = converter.to_value(value)
 
+        # Set the attribute's new value
         object.__setattr__(self, name, value)
 
-        if mapper:
-            log.critical(mapper.attrs)
-
-        if mapper and name in mapper.attrs:
-            if mapper.auto:
-                self.yorm_mapper.store()
-            else:
-                log.trace("automatic storage is off")
-
-    def __enter__(self):
-        log.debug("turning off automatic storage...")
-        mapper = getattr(self, MAPPER)
-        mapper.auto = False
-
-    def __exit__(self, *_):
-        log.debug("turning on automatic storage...")
-        mapper = getattr(self, MAPPER)
-        mapper.store()
+        # Store the attribute to disk
+        if mapper and mapper.auto and name in mapper.attrs:
+            self.yorm_mapper.store()